requirements.txt
# Dépendances pipeline Data Engineering (Airflow workers + scripts locaux)

# HTTP & téléchargement
requests==2.31.0
urllib3==2.0.7

# Données
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2
openpyxl==3.1.5
xlrd==2.0.1

# Base de données
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Stockage objet MinIO/S3
boto3==1.34.14
botocore==1.34.14

# Géospatial
geopandas==0.14.1
shapely==2.0.2
fiona==1.9.5
pyproj==3.6.1

# Qualité des données
great-expectations==0.18.22

# dbt
dbt-core==1.8.0
dbt-postgres==1.8.0

# OpenLineage
openlineage-python==1.9.0
openlineage-airflow==1.9.0
openlineage-dbt==1.9.0

# Monitoring
prometheus-client==0.19.0

# Secrets
hvac==2.1.0

# Utilitaires
python-dotenv==1.0.0
click==8.1.7
pydantic==2.5.3
tenacity==8.2.3           
tqdm==4.66.1              
loguru==0.7.2            
